Manuscript Published: Updated citation information & code for Fig. 5
Our manuscript has been published, so up to date information on where it can be accessed and how to cite it has been added to the README file.

During peer review, a reviewer suggested we change split Fig. 5 into 2 sub-figures. The script to reproduce Fig. 5 has been updated to produce the updated version of this figure as well as the original.